spotifyotd/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
dnspython==2.2.1
idna==3.4
imbalanced-learn==0.9.1
joblib==1.2.0
numpy==1.23.5
pandas==1.5.2
Pillow==9.3.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.6
urllib3==1.26.13
